To bug-fikser fra beta-tester: desimal-komma bevares + uke-kort-raden fjernet
Bug 1 — desimaler trunkert (7,87 km → 7,0): parseFloat("7,87") stopper
på norsk komma og ga 7 ved lagring. Ny felles parseDecimal-helper
(lib/parse-decimal.ts) normaliserer komma→punktum og er byttet inn på
alle 51 parse-steder for brukerinput: distanse, laktat, styrkevekt,
sekke-/pulkvekt, stigning, HRV, søvn, kroppsvekt, pace, ernæring,
utstyrspriser, ski-test-temperaturer, mal-/plan-felter m.m.
DB-kolonnene var allerede numeric(x,2) — ingen skjemaendring. Data
lagret uten desimaler før fiksen er tapt; nye lagres riktig.
Watt/høydemeter/puls er bevisst heltall i DB og er uendret.

Bug 2 — uke-kort-raden på Dagbok («Denne uken / Km / Økter») viste
tomme verdier: fjernet helt (redundant — Uke-toggle i kalenderen og
startsiden dekker det). Tilhørende uke-query og summeringslogikk i
DagbokPageView er ryddet bort — ett DB-kall mindre per sidevisning.

+++ app/actions/health.ts
@@ -3,6 +3,7 @@
 import { revalidatePath } from 'next/cache'
 import { createClient } from '@/lib/supabase/server'
 import { resolveTargetUser } from '@/lib/target-user'
+import { parseDecimal } from '@/lib/parse-decimal'
 
 export async function saveDailyHealth(data: {
   date: string
@@ -22,10 +23,10 @@ export async function saveDailyHealth(data: {
     user_id: resolved.userId,
     date: data.date,
     resting_hr: parseInt(data.resting_hr) || null,
-    hrv_ms: parseFloat(data.hrv_ms) || null,
-    sleep_hours: parseFloat(data.sleep_hours) || null,
+    hrv_ms: parseDecimal(data.hrv_ms) || null,
+    sleep_hours: parseDecimal(data.sleep_hours) || null,
     sleep_quality: data.sleep_quality,
-    body_weight_kg: parseFloat(data.body_weight_kg) || null,
+    body_weight_kg: parseDecimal(data.body_weight_kg) || null,
     notes: data.notes || null,
     updated_at: new Date().toISOString(),
   }
